# java-test-api/target/classes/api/openapi.yml
openapi: 3.0.2
info:
  title: Java test
  description: "Java test for Inditex"
  termsOfService: http://swagger.io/terms/
  contact:
    email: javiernp19@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1-SNAPSHOT
servers:
- url: 'http://localhost:8080/'

paths:
  /price:
    post:
      summary: Get a list of User with their descriptions.
      operationId: getPriceByFilter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Price'
                          
components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string 
          
    PriceFilter:
      type: object
      properties:
        date:
          type: string
          format: date-time
        product-id:
          type: integer
          format: int64
        brand-id:
          type: integer
          format: int64
          
    Brand:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
  
    Price:
      type: object
      properties:
        id:
          type: integer
          format: int64
        brand:
          $ref: '#/components/schemas/Brand'
        product:
          $ref: '#/components/schemas/Product'    
        start-date:
          type: string
          format: date-time
        end-date:
          type: string
          format: date-time
        price:
          type: number
          format: double
        curr:
          type: string
          